Intent: Consolidate utility modules under tnfr.utils Operators involved: [Transition, Recursivity] Affected invariants: [#8] Key changes: - Introduced the tnfr.utils package exposing logging/import, data, graph and JSON helpers - Replaced existing imports to consume tnfr.utils APIs and added deprecation shims for legacy modules - Updated tests and docs to reference the new utility package while preserving logging behaviour Expected risks/dissonances: Deprecation warnings during migration; monitored cached-import logging to ensure stability Metrics: C(t), Si, νf unchanged; logging behaviour verified via tests Equivalence map: tnfr.import_utils -> tnfr.utils.init, tnfr.logging_utils -> tnfr.utils.init, tnfr.collections_utils/value_utils/graph_utils/json_utils -> tnfr.utils.*
