+++ src/tnfr/cli/execution.py
@@ -29,9 +29,8 @@ from ..config import apply_config
 from ..io import read_structured_file, safe_write, StructuredFileError
 from ..glyph_history import ensure_history
 from ..ontosim import preparar_red
-from ..logging_utils import get_logger
 from ..types import Glyph
-from ..json_utils import json_dumps
+from ..utils import get_logger, json_dumps
 from ..flatten import parse_program_tokens
 
 from .arguments import _args_to_dict
