+++ src/tnfr/glyph_history.py
@@ -9,8 +9,7 @@ from collections.abc import Iterable, Mapping
 from functools import lru_cache
 
 from .constants import get_param
-from .collections_utils import ensure_collection
-from .logging_utils import get_logger
+from .utils import ensure_collection, get_logger
 
 logger = get_logger(__name__)
