+++ src/tnfr/metrics/coherence.py
@@ -14,7 +14,6 @@ from ..constants import (
 from ..callback_utils import CallbackEvent, callback_manager
 from ..glyph_history import ensure_history, append_metric
 from ..alias import collect_attr, get_attr, set_attr
-from ..collections_utils import normalize_weights
 from ..helpers.numeric import clamp01
 from ..cache import ensure_node_index_map
 from .common import compute_coherence, min_max_range
@@ -27,8 +26,7 @@ from ..observers import (
     phase_sync,
 )
 from ..sense import sigma_vector
-from ..import_utils import get_numpy
-from ..logging_utils import get_logger
+from ..utils import get_logger, get_numpy, normalize_weights
 
 logger = get_logger(__name__)
