+++ src/tnfr/metrics/export.py
@@ -6,11 +6,11 @@ import csv
 import math
 from itertools import zip_longest, tee
 
+from ..constants_glyphs import GLYPHS_CANONICAL
 from ..glyph_history import ensure_history
 from ..io import safe_write
-from ..constants_glyphs import GLYPHS_CANONICAL
+from ..utils import json_dumps
 from .core import glyphogram_series
-from ..json_utils import json_dumps
 
 
 def _write_csv(path, headers, rows):
