+++ src/tnfr/metrics/trig_cache.py
@@ -11,10 +11,10 @@ from dataclasses import dataclass
 from typing import Any, Iterable, Mapping
 
 from ..alias import get_attr
-from ..constants import get_aliases
 from ..cache import edge_version_cache
-from ..import_utils import get_numpy
+from ..constants import get_aliases
 from ..types import GraphLike
+from ..utils import get_numpy
 
 ALIAS_THETA = get_aliases("THETA")
