+++ src/tnfr/selector.py
@@ -18,7 +18,7 @@ from .constants import DEFAULTS
 from .constants.core import SELECTOR_THRESHOLD_DEFAULTS
 from .helpers.numeric import clamp01
 from .metrics.common import compute_dnfr_accel_max
-from .collections_utils import is_non_string_sequence
+from .utils import is_non_string_sequence
 
 
 HYSTERESIS_GLYPHS: set[str] = {"IL", "OZ", "ZHIR", "THOL", "NAV", "RA"}
