+++ tests/test_compute_Si_numpy_usage.py
@@ -1,9 +1,9 @@
 import math
 
+from tnfr.alias import set_attr
 from tnfr.constants import get_aliases
 from tnfr.metrics.sense_index import compute_Si
-from tnfr.alias import set_attr
-import tnfr.import_utils as import_utils
+import tnfr.utils.init as utils_init
 
 ALIAS_THETA = get_aliases("THETA")
 ALIAS_VF = get_aliases("VF")
@@ -33,7 +33,7 @@ def test_compute_Si_uses_module_numpy_and_propagates(monkeypatch, graph_canon):
         return 0.0
 
     monkeypatch.setattr(
-        import_utils,
+        utils_init,
         "cached_import",
         lambda module, attr=None, **kwargs: sentinel if module == "numpy" else None,
     )
