+++ tests/test_logging_threadsafe.py
@@ -2,7 +2,7 @@ from concurrent.futures import ThreadPoolExecutor
 import importlib
 import logging
 
-import tnfr.logging_utils as logging_utils
+import tnfr.utils.init as logging_utils
 
 
 def reload_logging_utils():
